+++ src/views/WorkoutPlansList.vue
@@ -3,6 +3,7 @@ import WorkoutPlansTable from "../components/WorkoutPlansTable.vue";
 import PageHeader from "../components/PageHeader.vue";
 import WorkoutPlansDataService from "../services/WorkoutPlansDataService";
 import WidgetAddButton from "../components/WidgetAddButton.vue";
+import Swal from "sweetalert2";
 </script>
 
 <script>
@@ -45,15 +46,28 @@ export default {
     },
 
     deleteWorkoutPlan(workoutPlanId, index) {
-      WorkoutPlansDataService.delete(workoutPlanId)
-        .then((response) => {
-          //console.log(response.data);
-          this.items.splice(index, 1);
-          this.page.message = this.items.length == 0 ? "No records found." : "";
-        })
-        .catch((error) => {
-          console.log(error);
-        });
+      Swal.fire({
+        title: "DELECTION CONFIRMATION!",
+        text: "Are you sure that you want to delete the item?",
+        icon: "warning",
+        confirmButtonText: "Yes, delete it.",
+        cancelButtonText: "No, cancel.",
+        showCancelButton: true,
+        showCloseButton: true,
+      }).then((result) => {
+        if (result.isConfirmed) {
+          WorkoutPlansDataService.delete(workoutPlanId)
+            .then((response) => {
+              //console.log(response.data);
+              this.items.splice(index, 1);
+              this.page.message =
+                this.items.length == 0 ? "No records found." : "";
+            })
+            .catch((error) => {
+              console.log(error);
+            });
+        }
+      });
     },
   },
 };
